requirements.txt
streamlit
pypdf2
python-dotenv
google-generativeai
pandas
plotly
